eduguardai/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0
bcrypt==4.1.2
python-dateutil==2.8.2
mysql-connector-python==8.2.0

